+++ src/gh_profiler/gh_profiler.py
@@ -4,7 +4,6 @@ The goal is to help make quick, evidence-based decisions about how much time
 to invest in reviewing PRs, and general interaction on open source projects.
 """
 
-import sys
 
 from .utils import profile_utils
 from .utils import analysis_utils